Fix failing tests due to float/int conversion
With a newer version of pytorch, adding or multiplying a python float
literal to an int tensor results in an int tensor. Before it resulted in
a float tensor, which the tests relied on. I fixed this by explictily
casting the integer tensors to floats.

+++ uwnet/test_constraints.py
@@ -31,7 +31,7 @@ def test_apply_linear_constraint():
 
 
 def test_fix_negative_moisture():
-    q = torch.arange(-4, 20) + .1
+    q = torch.arange(-4, 20).float() + .1
     layer_mass = torch.rand(len(q)) + .1
 
     q_new = fix_negative_moisture(q, layer_mass)
